backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Image processing
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
scipy==1.11.4
scikit-image==0.22.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
Werkzeug==3.0.1
